refactor: cleanup unused legacy utils
