+++ packages/blocks/src/code-block/code-block.ts
@@ -28,10 +28,10 @@ import { createLitPortal } from '../_common/components/portal.js';
 import { bindContainerHotkey } from '../_common/components/rich-text/keymap/index.js';
 import type { RichText } from '../_common/components/rich-text/rich-text.js';
 import { ArrowDownIcon } from '../_common/icons/index.js';
+import { listenToThemeChange } from '../_common/theme/utils.js';
 import { PAGE_HEADER_HEIGHT } from '../_legacy/consts.js';
 import { queryCurrentMode } from '../_legacy/index.js';
 import { getService } from '../_legacy/service/index.js';
-import { listenToThemeChange } from '../_legacy/theme/utils.js';
 import type { CodeBlockModel } from './code-model.js';
 import { CodeOptionTemplate } from './components/code-option.js';
 import { LangList } from './components/lang-list.js';
