+++ packages/blocks/src/page-block/edgeless/components/component-toolbar/change-brush-button.ts
@@ -9,7 +9,7 @@ import { css, html, LitElement } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import type { CssVariableName } from '../../../../_legacy/theme/css-variables.js';
+import type { CssVariableName } from '../../../../_common/theme/css-variables.js';
 import { LineWidth } from '../../../../_legacy/utils/types.js';
 import type {
   BrushElement,
