+++ packages/blocks/src/page-block/edgeless/components/component-toolbar/change-connector-button.ts
@@ -20,7 +20,7 @@ import {
   SmallArrowDownIcon,
   StraightLineIcon,
 } from '../../../../_common/icons/index.js';
-import type { CssVariableName } from '../../../../_legacy/theme/css-variables.js';
+import type { CssVariableName } from '../../../../_common/theme/css-variables.js';
 import { LineWidth } from '../../../../_legacy/utils/types.js';
 import type { PhasorElementType } from '../../../../surface-block/index.js';
 import {
