+++ packages/blocks/src/page-block/edgeless/components/component-toolbar/change-text-menu.ts
@@ -16,7 +16,7 @@ import {
   TextAlignLeftIcon,
   TextAlignRightIcon,
 } from '../../../../_common/icons/index.js';
-import type { CssVariableName } from '../../../../_legacy/theme/css-variables.js';
+import type { CssVariableName } from '../../../../_common/theme/css-variables.js';
 import {
   Bound,
   normalizeShapeBound,
