+++ packages/blocks/src/page-block/edgeless/edgeless-page-block.ts
@@ -15,6 +15,7 @@ import { customElement, query, state } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
 import { toast } from '../../_common/components/toast.js';
+import { listenToThemeChange } from '../../_common/theme/utils.js';
 import { EdgelessClipboard } from '../../_legacy/clipboard/index.js';
 import { BLOCK_ID_ATTR } from '../../_legacy/consts.js';
 import type { EdgelessTool, Point, Selectable } from '../../_legacy/index.js';
@@ -25,7 +26,6 @@ import {
   type TopLevelBlockModel,
 } from '../../_legacy/index.js';
 import { getService } from '../../_legacy/service/index.js';
-import { listenToThemeChange } from '../../_legacy/theme/utils.js';
 import type { ImageBlockModel } from '../../image-block/index.js';
 import type { NoteBlockModel } from '../../note-block/index.js';
 import { EdgelessBlockType } from '../../surface-block/edgeless-types.js';
