+++ packages/blocks/src/surface-block/surface-block.ts
@@ -9,6 +9,11 @@ import { css, html, nothing } from 'lit';
 import { customElement, query } from 'lit/decorators.js';
 
 import {
+  type CssVariableName,
+  isCssVariable,
+} from '../_common/theme/css-variables.js';
+import { getThemePropertyValue } from '../_common/theme/utils.js';
+import {
   bringForward,
   type EdgelessElement,
   reorder,
@@ -19,11 +24,6 @@ import {
   sendBackward,
   type TopLevelBlockModel,
 } from '../_legacy/index.js';
-import {
-  type CssVariableName,
-  isCssVariable,
-} from '../_legacy/theme/css-variables.js';
-import { getThemePropertyValue } from '../_legacy/theme/utils.js';
 import { EdgelessConnectorManager } from '../page-block/edgeless/connector-manager.js';
 import type { EdgelessPageBlockComponent } from '../page-block/edgeless/edgeless-page-block.js';
 import { EdgelessFrameManager } from '../page-block/edgeless/frame-manager.js';
