+++ tests/utils/asserts.ts
@@ -21,7 +21,6 @@ import type {
   PageBlockModel,
 } from '../../packages/blocks/src/index.js';
 import { assertExists } from '../../packages/global/src/utils.js';
-import { toHex } from '../../packages/global/src/utils.js';
 import type { BlockElement } from '../../packages/lit/src/index.js';
 import type { RichText } from '../../packages/playground/examples/virgo/test-page.js';
 import {
@@ -845,6 +844,31 @@ export async function assertEdgelessNoteBackground(
     });
 
   expect(backgroundColor).toEqual(`var(${color})`);
+}
+
+function toHex(color: string) {
+  let r, g, b;
+
+  if (color.startsWith('#')) {
+    color = color.substr(1);
+    if (color.length === 3) {
+      color = color.replace(/./g, '$&$&');
+    }
+    [r, g, b] = color.match(/.{2}/g)?.map(hex => parseInt(hex, 16)) ?? [];
+  } else if (color.startsWith('rgba')) {
+    [r, g, b] = color.match(/\d+/g)?.map(Number) ?? [];
+  } else if (color.startsWith('rgb')) {
+    [r, g, b] = color.match(/\d+/g)?.map(Number) ?? [];
+  } else {
+    throw new Error('Invalid color format');
+  }
+
+  if (r === undefined || g === undefined || b === undefined) {
+    throw new Error('Invalid color format');
+  }
+
+  const hex = ((r << 16) | (g << 8) | b).toString(16);
+  return '#' + '0'.repeat(6 - hex.length) + hex;
 }
 
 export async function assertEdgelessColorSameWithHexColor(
